refactor: presentation 패키지로 이동  // livedata 수정 필요
