+++ .idea/deploymentTargetDropDown.xml
@@ -10,12 +10,12 @@
               <deviceKey>
                 <Key>
                   <type value="VIRTUAL_DEVICE_PATH" />
-                  <value value="C:\Users\msgal\.android\avd\Pixel_8_API_33.avd" />
+                  <value value="C:\Users\msgal\.android\avd\Medium_Phone_API_33.avd" />
                 </Key>
               </deviceKey>
             </Target>
           </targetSelectedWithDropDown>
-          <timeTargetWasSelectedWithDropDown value="2024-06-03T09:26:41.768321600Z" />
+          <timeTargetWasSelectedWithDropDown value="2024-06-03T11:37:51.081808200Z" />
         </State>
       </entry>
     </value>
