+++ app/src/main/AndroidManifest.xml
@@ -25,8 +25,8 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name=".consumption.ConsumptionActivity"/>
-        <activity android:name=".consumption.CalculationActivity"/>
+        <activity android:name=".consumption.presentation.ConsumptionActivity"/>
+        <activity android:name=".consumption.presentation.CalculationActivity"/>
     </application>
 
 </manifest>
