+++ app/src/main/java/com/example/donotlate/consumption/presentation/CalculationActivity.kt
@@ -1,4 +1,4 @@
-package com.example.donotlate.consumption
+package com.example.donotlate.consumption.presentation
 
 import android.os.Bundle
 import androidx.activity.enableEdgeToEdge
