+++ app/src/main/java/com/example/donotlate/consumption/presentation/CalculationFragment1.kt
@@ -1,10 +1,9 @@
-package com.example.donotlate.consumption
+package com.example.donotlate.consumption.presentation
 
 import android.annotation.SuppressLint
 import android.app.DatePickerDialog
 import android.graphics.Color
 import android.os.Bundle
-import android.util.Log
 import android.view.LayoutInflater
 import android.view.MotionEvent
 import android.view.View
