+++ app/src/main/java/com/example/donotlate/consumption/presentation/CalculationFragment2.kt
@@ -1,8 +1,7 @@
-package com.example.donotlate.consumption
+package com.example.donotlate.consumption.presentation
 
 import android.annotation.SuppressLint
 import android.os.Bundle
-import android.util.Log
 import android.view.LayoutInflater
 import android.view.MotionEvent
 import android.view.View
