+++ app/src/main/java/com/example/donotlate/consumption/presentation/ConsumptionModel.kt
@@ -1,4 +1,4 @@
-package com.example.donotlate.consumption
+package com.example.donotlate.consumption.presentation
 
 import android.os.Parcelable
 import kotlinx.parcelize.Parcelize
